created slider, but need a programmer's assistance
I created the slider component and put in the Fill color. I made it the values min 0 and max 5. For some reason, when I am sliding it, the bar goes way out of the fill range and outside of the bar when it goes to 5. I am going to need help with scripting this please.